knex fixed completely

+++ src/config/migrations/20251020083458_alter-shops.ts
@@ -1,19 +1,27 @@
 import type { Knex } from "knex";
+import type { IIndexRow } from "../../types/knex.js";
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.alterTable('shops', table=> {
-        table.unique('user_id', {
-            indexName: 'uq_shops_user_id'
-        });
-        
-        table
-            .foreign('user_id')
-            .references('ID')
-            .inTable('users')
-            .onDelete('CASCADE')
-            .onUpdate('CASCADE')
-            .withKeyName('fk_shops_user_id')
-    })
+    const result = await knex.raw(`
+        SHOW INDEX FROM shops WHERE Key_name = 'uq_shops_user_id';
+    `);
+    
+    const existingIndex: IIndexRow[] = result[0];
+    if (existingIndex.length === 0) {
+        await knex.schema.alterTable('shops', table=> {
+            table
+                .foreign('user_id')
+                .references('ID')
+                .inTable('users')
+                .onDelete('CASCADE')
+                .onUpdate('CASCADE')
+                .withKeyName('fk_shops_user_id')
+    
+            table.unique('user_id', {
+                indexName: 'uq_shops_user_id'
+            });
+        })
+    }
 }
 
 export async function down(knex: Knex): Promise<void> {
