+++ src/config/migrations/20251020083801_alter-sections.ts
@@ -1,17 +1,24 @@
 import type { Knex } from "knex";
+import type { IIndexRow } from "../../types/knex.js";
 
 
 export async function up(knex: Knex): Promise<void> {
-    await knex.schema.alterTable('sections', table=> {
+    const result = await knex.raw(`
+        SHOW INDEX FROM sections WHERE Key_name IN ('fk_sections_material_id')        
+    `);
+    const existingIndex: IIndexRow[] = result[0];
 
-        table
-            .foreign('material_id')
-            .references('ID')
-            .inTable('materials')
-            .onDelete('CASCADE')
-            .onUpdate('CASCADE')
-            .withKeyName('fk_sections_material_id')
-    })
+    if (!existingIndex) {
+        await knex.schema.alterTable('sections', table=> {
+            table
+                .foreign('material_id')
+                .references('ID')
+                .inTable('materials')
+                .onDelete('CASCADE')
+                .onUpdate('CASCADE')
+                .withKeyName('fk_sections_material_id')
+        })
+    }
 }
